+++ src/main/java/com/example/demo/model/dto/AppointmentFormDTO.java
@@ -1,5 +1,5 @@
 package com.example.demo.model.dto;
 
 public class AppointmentFormDTO {
-	private Integer AppointmentFormId;
+	private Long AppointmentFormId;
 }
