+++ src/main/java/com/example/demo/model/dto/FitnessInstructorDTO.java
@@ -7,7 +7,7 @@ import com.example.demo.model.entity.AppointmentForm;
 import com.example.demo.model.entity.Class;
 
 public class FitnessInstructorDTO {
-	private Integer FitnessInstructorId;		// *教練邊號
+	private Long FitnessInstructorId;		// *教練邊號
 	private String fitnessInstructorName;		// 教練姓名
 	
 	private List<Class> classes = new ArrayList<>();	// 授課類型
