+++ src/main/java/com/example/demo/model/dto/MemberDTO.java
@@ -10,7 +10,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class MemberDTO {
-	private Integer member_id;	// *會員邊號
+	private Long member_id;	// *會員邊號
 	private String mrmber_name;	// 會員名稱
 	
 	private ArrayList<ActivitySchedule> sing_up = new ArrayList<>();	// 已報名課程
