+++ src/main/java/com/example/demo/model/entity/Member.java
@@ -14,7 +14,7 @@ public class Member {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)	// 自動生成，從1開始，每次+1，過號不補
-	private Integer member_id;	// *會員邊號
+	private Long member_id;	// *會員邊號
 	
 	@Column(nullable = false)
 	private String mrmber_name;	// 會員名稱
